+++ src/main/java/chess/controller/ChessGameController.java
@@ -1,9 +1,9 @@
 package chess.controller;
 
 import chess.domain.ChessGame;
+import chess.domain.Command;
 import chess.domain.Team;
 import chess.exception.StartCommandException;
-import chess.view.Command;
 import chess.view.InputView;
 import chess.view.OutputView;
