+++ src/test/java/chess/view/CommandTest.java
@@ -1,5 +1,6 @@
 package chess.view;
 
+import chess.domain.Command;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
